third_party/WebKit/LayoutTests/fast/html/imports/import-style-quirks.html

<!-- quirks mode -->
<html>
<head>
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <link rel="import" href="resources/import-quirks.html">
  <link rel="import" href="resources/import-no-quirks.html">
</head>
<body>
  <div id="MyElement">MyElement</div>
  <div class="MyClass" id="OtherElement">OtherElement</div>
</body>
<script>
'use strict';

let master_my_el = document.getElementById('MyElement');
let master_my_class = document.getElementById('OtherElement');

test(() => {
  assert_equals(document.compatMode, 'BackCompat');
}, 'This document should be in quirks mode.');

test(() => {
  assert_equals(document.getElementById('myelement'), null);
  assert_equals(document.getElementById('MyElement'), master_my_el);
  assert_equals(document.getElementById('MYELEMENT'), null);
}, 'getElementById should be case sensitive regardless of mode.');

test(() => {
  assert_equals(document.getElementsByClassName('myclass')[0],
                master_my_class);
  assert_equals(document.getElementsByClassName('MyClass')[0],
                master_my_class);
  assert_equals(document.getElementsByClassName('MYCLASS')[0],
                master_my_class);
}, 'getElementsByClassName should be case insensitive in quirks mode.');

test(() => {
  assert_equals(document.querySelector('#myelement'), master_my_el);
  assert_equals(document.querySelector('#MyElement'), master_my_el);
  assert_equals(document.querySelector('#MYELEMENT'), master_my_el);

  assert_equals(document.querySelector('.myclass'), master_my_class);
  assert_equals(document.querySelector('.MyClass'), master_my_class);
  assert_equals(document.querySelector('.MYCLASS'), master_my_class);
}, 'querySelector should be case insensitive when in quirks mode.');
</script>
</html>
